fix(ui): lazy-load run inspector copy
