+++ scripts/control-ui-i18n-verify.ts
@@ -6,6 +6,7 @@ import { fileURLToPath, pathToFileURL } from "node:url";
 import {
   loadControlUiTranslationMemory,
   materializeControlUiLocaleCatalog,
+  mergeControlUiTranslationMaps,
 } from "./lib/control-ui-i18n-catalog.ts";
 import { CONTROL_UI_LOCALE_ENTRIES } from "./lib/control-ui-i18n-config.ts";
 import { syncControlUiRawCopyBaseline } from "./lib/control-ui-i18n-raw-copy.ts";
@@ -21,6 +22,7 @@ const ROOT = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
 const LOCALES_DIR = path.join(ROOT, "ui", "src", "i18n", "locales");
 const I18N_ASSETS_DIR = path.join(ROOT, "ui", "src", "i18n", ".i18n");
 const SOURCE_LOCALE_PATH = path.join(LOCALES_DIR, "en.ts");
+const ACTIVITY_SOURCE_LOCALE_PATH = path.join(LOCALES_DIR, "en-activity.ts");
 const FALLBACK_BASELINE_PATH = path.join(I18N_ASSETS_DIR, "catalog-fallbacks.json");
 const FALLBACK_BASELINE_VERSION = 1;
 
@@ -50,6 +52,26 @@ async function loadLocaleMap(filePath: string, exportName: string): Promise<Tran
   }
   const mod = await importLocaleModule<Record<string, TranslationMap>>(filePath);
   return mod[exportName] ?? null;
+}
+
+async function loadSourceLocaleMap(): Promise<TranslationMap> {
+  const source = await loadLocaleMap(SOURCE_LOCALE_PATH, "en");
+  const activitySource = (
+    await importLocaleModule<{
+      registerActivityEnglish: { catalog: TranslationMap };
+    }>(ACTIVITY_SOURCE_LOCALE_PATH)
+  ).registerActivityEnglish.catalog;
+  if (!source || !activitySource) {
+    throw new Error("Control UI English source catalogs are incomplete");
+  }
+  return mergeControlUiTranslationMaps(source, activitySource);
+}
+
+async function readSourceLocaleRaw(): Promise<string> {
+  const sources = await Promise.all(
+    [SOURCE_LOCALE_PATH, ACTIVITY_SOURCE_LOCALE_PATH].map((filePath) => readFile(filePath, "utf8")),
+  );
+  return sources.join("\n");
 }
 
 function extractPlaceholders(text: string): string[] {
@@ -136,11 +158,8 @@ async function buildCatalogFallbackBaseline(
     allowCatalogDrift?: boolean;
   } = {},
 ): Promise<CatalogFallbackBaseline> {
-  const sourceRaw = await readFile(SOURCE_LOCALE_PATH, "utf8");
-  const sourceMap = await loadLocaleMap(SOURCE_LOCALE_PATH, "en");
-  if (!sourceMap) {
-    throw new Error("ui/src/i18n/locales/en.ts does not export en");
-  }
+  const sourceRaw = await readSourceLocaleRaw();
+  const sourceMap = await loadSourceLocaleMap();
   const sourceFlat = flattenControlUiCatalog(sourceMap, "en");
   const localeFlats = new Map<string, Map<string, string>>();
   for (const entry of CONTROL_UI_LOCALE_ENTRIES) {
@@ -191,10 +210,7 @@ function printCatalogFallbackSummary(baseline: CatalogFallbackBaseline) {
 }
 
 async function verifyControlUiSourceCatalogShape() {
-  const sourceMap = await loadLocaleMap(SOURCE_LOCALE_PATH, "en");
-  if (!sourceMap) {
-    throw new Error("ui/src/i18n/locales/en.ts does not export en");
-  }
+  const sourceMap = await loadSourceLocaleMap();
   const sourceFlat = flattenControlUiCatalog(sourceMap, "en");
   process.stdout.write(`control-ui-i18n: source: keys=${sourceFlat.size}\n`);
 }
