+++ ui/config/control-ui-locales.ts
@@ -4,9 +4,11 @@ import type { Plugin } from "vite";
 import {
   loadControlUiTranslationMemory,
   materializeControlUiLocaleCatalog,
+  mergeControlUiTranslationMaps,
 } from "../../scripts/lib/control-ui-i18n-catalog.ts";
 import { CONTROL_UI_LOCALE_ENTRIES } from "../../scripts/lib/control-ui-i18n-config.ts";
 import { flattenTranslations } from "../../scripts/lib/control-ui-i18n-sync-plan.ts";
+import { registerActivityEnglish } from "../src/i18n/locales/en-activity.ts";
 import { en } from "../src/i18n/locales/en.ts";
 
 const localeModulePrefix = "virtual:openclaw-control-ui-locale/";
@@ -17,6 +19,7 @@ const i18nAssetsDir = path.resolve(
   "../src/i18n/.i18n",
 );
 const locales = new Set(CONTROL_UI_LOCALE_ENTRIES.map(({ locale }) => locale));
+const sourceCatalog = mergeControlUiTranslationMaps(en, registerActivityEnglish.catalog);
 
 export function controlUiLocaleModulesPlugin(): Plugin {
   return {
@@ -42,7 +45,7 @@ export function controlUiLocaleModulesPlugin(): Plugin {
       if (memory.size === 0) {
         throw new Error(`Control UI ${locale} translation memory is missing or empty`);
       }
-      const catalog = materializeControlUiLocaleCatalog(flattenTranslations(en), memory);
+      const catalog = materializeControlUiLocaleCatalog(flattenTranslations(sourceCatalog), memory);
       return `export default ${JSON.stringify(catalog)};`;
     },
   };
